+++ bin/.gitignore
@@ -1,3 +1,11 @@
 /main/
 /test/
 /ui/
+/deprecatedui/
+/factory/
+/products/
+/ingredient/
+/abstractclases/
+/store/
+/lasagna/
+/pizza/
