--- src/main/java/p1/LasagnaType.java
@@ -1,5 +1,0 @@
-package main.java.p1;
-
-public enum LasagnaType {
-STANDAR
-}
