+++ src/pizza/GreekPizzaBranchB.java
@@ -1,24 +1,34 @@
-package main.java.p1;
+package pizza;
 
-public class GreekPizzaBranchA extends AbstractPizza {
+import ingredient.IngredientType;
+import ingredient.Ingredients;
+
+public class GreekPizzaBranchB extends AbstractPizza {
     private static final double PIZZA_COST_BASE = 10;
-    Ingredients thick_dough;
+    Ingredients thin_dough;
     Ingredients greek_cheese;
     Ingredients olive;
+    Ingredients asparagus;
     Ingredients tomato_slices;
     
 
-    GreekPizzaBranchA() {
+    public GreekPizzaBranchB() {
         super(PizzaTypes.GREEK);
         pizzaCost = PIZZA_COST_BASE;
-        thick_dough = new Ingredients(IngredientType.THICK_DOUGH);
+        thin_dough = new Ingredients(IngredientType.THIN_DOUGH);
         greek_cheese = new Ingredients(IngredientType.GREEK_CHEESE);
         olive = new Ingredients(IngredientType.OLIVE);
         tomato_slices = new Ingredients(IngredientType.TOMATO_SLICES);
+        asparagus = new Ingredients(IngredientType.ASPARAGUS);
+       
 
-        ingredientsList.add(thick_dough);
+        ingredientsList.add(thin_dough);
         ingredientsList.add(greek_cheese);
         ingredientsList.add(olive);
         ingredientsList.add(tomato_slices);
+        ingredientsList.add(asparagus);
     }
+
+   
+
 }
