+++ src/pizza/Lasagna.java
@@ -1,6 +1,9 @@
-package main.java.p1;
+package pizza;
 
 import java.util.HashSet;
+
+import abstractclases.Product;
+import ingredient.IngredientsLasagna;
 
 public abstract class Lasagna extends Product {
