+++ src/pizza/LasagnaBranchB.java
@@ -1,4 +1,7 @@
-package main.java.p1;
+package pizza;
+
+import ingredient.IngredientType;
+import ingredient.IngredientsLasagna;
 
 public class LasagnaBranchB extends Lasagna {
     private static final double LASAGNA_COST_BASE = 6;
