+++ src/pizza/LasagnaType.java
@@ -1,0 +1,6 @@
+package pizza;
+
+public enum LasagnaType {
+	PICK,
+	STANDAR
+}
