+++ src/test/java/Mani.java
@@ -1,4 +1,11 @@
-package main.java.p1;
+package test.java;
+
+import factory.FactoryBranchA;
+import ingredient.IngredientType;
+import ingredient.Ingredients;
+import pizza.AbstractPizza;
+import pizza.PizzaTypes;
+import store.Order;
 
 public class Mani {
